# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Features Service API",
    "version": "3.0.0",
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  },
  "paths": {
...
    "/features/{dynastySlug}/inputs": {
      "get": {
        "summary": "Get input definitions by dynasty slug",
        "description": "Returns the input field definitions for the **active version** of a dynasty. The path param must be a **dynasty slug** (e.g. `sales-cold-email`), NOT a versioned slug.\n\nReturns 404 if no active feature exists for this dynasty slug.\n\nThe response includes both the resolved `slug` (versioned) and the `dynastySlug` for clarity.",
        "tags": [
          "Features"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Dynasty slug (e.g. 'sales-cold-email'). Must be a dynasty slug — versioned slugs will 404."
            },
            "required": true,
            "description": "Dynasty slug (e.g. 'sales-cold-email'). Must be a dynasty slug — versioned slugs will 404.",
            "name": "dynastySlug",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal org UUID from client-service"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal user UUID from client-service"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Run ID for tracking and billing"
            },
            "required": true,
            "description": "Run ID for tracking and billing",
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Feature inputs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "slug": {
                      "type": "string",
                      "description": "Resolved versioned slug (e.g. 'sales-cold-email-v2')"
                    },
                    "dynastySlug": {
                      "type": "string",
                      "description": "Dynasty slug used in the request (e.g. 'sales-cold-email')"
                    },
                    "name": {
                      "type": "string",
                      "description": "Dynasty display name"
                    },
                    "inputs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "key": {
                            "type": "string",
                            "minLength": 1
                          },
                          "label": {
                            "type": "string",
                            "minLength": 1
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "text",
                              "textarea",
                              "number",
                              "select"
                            ]
                          },
                          "placeholder": {
                            "type": "string",
                            "minLength": 1
                          },
                          "description": {
                            "type": "string",
                            "minLength": 1
                          },
                          "extractKey": {
                            "type": "string",
                            "minLength": 1
                          },
                          "options": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        },
                        "required": [
                          "key",
                          "label",
                          "type",
                          "placeholder",
                          "description",
                          "extractKey"
                        ]
                      },
                      "description": "Input definitions."
                    }
                  },
                  "required": [
                    "slug",
                    "dynastySlug",
                    "name",
                    "inputs"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No active feature for this dynasty slug",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/features/dynasty": {
      "get": {
        "summary": "Resolve dynasty identity from a versioned feature slug",
        "description": "Returns the stable, unversioned dynasty identifiers for a given feature slug. Used by workflow-service to compose workflow names.\n\nExample: `?slug=sales-cold-email-sophia-v2` → `{ feature_dynasty_name: 'Sales Cold Email Sophia', feature_dynasty_slug: 'sales-cold-email-sophia' }`\n\nWorks for both active and deprecated features.",
        "tags": [
          "Features"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "The versioned feature slug to resolve"
            },
            "required": true,
            "description": "The versioned feature slug to resolve",
            "name": "slug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal org UUID from client-service"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Internal user UUID from client-service"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Run ID for tracking and billing"
            },
            "required": true,
            "description": "Run ID for tracking and billing",
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Dynasty identity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "feature_dynasty_name": {
                      "type": "string",
                      "description": "Stable dynasty name (unversioned), e.g. 'Sales Cold Email Sophia'"
                    },
                    "feature_dynasty_slug": {
                      "type": "string",
                      "description": "Stable dynasty slug (unversioned), e.g. 'sales-cold-email-sophia'"
                    }
                  },
                  "required": [
                    "feature_dynasty_name",
                    "feature_dynasty_slug"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing slug parameter",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Feature not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}